+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Update to Slim 4
 - Update to Babel 7
 - Refactored frontend to use a runtime-only build
+- User admin now also displays "incoming" characters that have been moved from another account.
 - Some UI and performance improvements
 - Fix: token state flag for SSOv2 tokens without scopes.